Move EphemeralEventListener to same file as agent
